+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.2
 
-* add example.
+* updae package.